Sync log detail when panel becomes visible
When the panel is made visible, check to see if there's a selected row or if it's different from the last log entry. This allows the detail panel to be closed when the user is selecting rows and have the selected row appear in the detail when the panel is opened.