Fix thread bug with raw rows
This fixes an issue that occurs when you first view a domain (which creates a binding to the raw rows collection) and then attempt an import (which was loading raw rows for the method table). Because import is async, this threw due to the binding. Fixed by loading a private copy of the methods to use for the import.